+++ Cargo.toml
@@ -1,14 +1,20 @@
 [package]
 name = "ferrux_canvas"
 description = "Abstraction tool to ease the use of the Pixels buffer"
-version = "0.1.0"
+version = "0.1.1"
 authors = ["Soto Estévez <ricardo@sotoestevez.dev>"]
 edition = "2021"
 readme = "./README.md"
 repository = "https://github.com/kriogenia/ferrux_canvas"
 license = "MIT OR Apache-2.0"
-keywords = [ "canvas", "pixels", "graphics" ]
-categories = [ "graphics" ]
+keywords = [ "canvas", "pixels", "graphics", "2D" ]
+categories = [ "graphics", "rendering" ]
+include = [
+    "/Cargo.*",
+    "/LICENSE-*",
+    "/README.md",
+    "/src/**/*",
+]
 
 [features]
 optimize = ["log/release_max_level_warn"]
