+++ examples/winit_example/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "winit_example"
+version = "0.1.0"
+authors = ["Ricardo Soto Estévez <ricardo@sotoestevez.dev>"]
+edition = "2021"
+publish = false
+
+[dependencies]
+ferrux_canvas = { path = "../.." }
+winit = "0.26"
