+++ examples/winit_example/README.md
@@ -1,0 +1,10 @@
+# Winit example
+
+Example for the Ferrux Canvas using **Winit**. It builds a new canvas with the default window
+and uses winit **EventLoop** to draw a morphing triangle using only the functions of the library.
+
+## Execution
+
+```shell
+cargo run --release --package winit_example
+```
